move logging from settings to codex/logger
