+++ codex/logger_base.py
@@ -1,5 +1,5 @@
 """Class to run librarian tasks inline without a thread."""
-from codex.settings.logging import get_logger
+from codex.logger.logging import get_logger
 
 
 class LoggerBaseMixin:
