+++ codex/run.py
@@ -12,8 +12,8 @@ from codex.asgi import application
 from codex.integrity import has_unapplied_migrations, rebuild_db, repair_db
 from codex.librarian.janitor.janitor import Janitor
 from codex.librarian.queue_mp import LIBRARIAN_QUEUE
+from codex.logger.logging import get_logger
 from codex.logger.mp_queue import LOG_QUEUE
-from codex.settings.logging import get_logger
 from codex.settings.settings import BACKUP_DB_PATH, HYPERCORN_CONFIG
 from codex.signals.os_signals import RESTART_EVENT, SHUTDOWN_EVENT, bind_signals
 from codex.version import VERSION
