+++ codex/views/admin/group.py
@@ -6,8 +6,8 @@ from rest_framework.viewsets import ModelViewSet
 
 from codex.librarian.notifier.tasks import LIBRARY_CHANGED_TASK
 from codex.librarian.queue_mp import LIBRARIAN_QUEUE
+from codex.logger.logging import get_logger
 from codex.serializers.admin import GroupSerializer
-from codex.settings.logging import get_logger
 
 
 LOG = get_logger(__name__)
