+++ codex/views/admin/user.py
@@ -11,8 +11,8 @@ from rest_framework.viewsets import ModelViewSet
 
 from codex.librarian.notifier.tasks import LIBRARY_CHANGED_TASK
 from codex.librarian.queue_mp import LIBRARIAN_QUEUE
+from codex.logger.logging import get_logger
 from codex.serializers.admin import UserChangePasswordSerializer, UserSerializer
-from codex.settings.logging import get_logger
 
 
 LOG = get_logger(__name__)
