+++ codex/views/browser/filters/search.py
@@ -4,7 +4,7 @@ from haystack.query import SearchQuerySet
 
 from codex.librarian.queue_mp import LIBRARIAN_QUEUE
 from codex.librarian.search.tasks import SearchIndexUpdateTask
-from codex.settings.logging import get_logger
+from codex.logger.logging import get_logger
 
 
 LOG = get_logger(__name__)
