+++ module_test.py
@@ -3,7 +3,7 @@ import sys
 
 from codex.librarian.covers.coverd import CoverCreator
 from codex.librarian.queue_mp import LIBRARIAN_QUEUE
-from codex.logger.log_queue import LOG_QUEUE
+from codex.logger.mp_queue import LOG_QUEUE
 from codex.models import Comic
