+++ tests/test_websocket.py
@@ -5,7 +5,7 @@ from codex.consumers.notifier import Channels
 from codex.librarian.notifier.notifierd import NotifierThread
 from codex.librarian.notifier.tasks import NotifierTask
 from codex.librarian.queue_mp import LIBRARIAN_QUEUE
-from codex.logger.log_queue import LOG_QUEUE
+from codex.logger.mp_queue import LOG_QUEUE
 
 
 pytestmark = pytest.mark.asyncio
